pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-flowcontrol"
version = "0.0.1"
description = "Design workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["django-treebeard", "python-dateutil", "django>=5.2"]

[dependency-groups]
dev = [
    "coverage>=7.6.12",
    "djlint>=1.36.1",
    "pytest>=8.3.4",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.9.0",
    "tox>=4.24.1",
]

[tool.hatch.build.targets.sdist]
packages = ["flowcontrol"]

[tool.hatch.build.targets.wheel]
packages = ["flowcontrol"]

[tool.ruff]
exclude = ["migrations", "build", "node_modules"]

[tool.ruff.lint]
ignore = ["E501", "C901"]
select = ["C", "E", "F", "W", "B", "I001"]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]
default-section = "third-party"

[tool.ruff.lint.isort.sections]
django = ["django"]

[tool.coverage.report]
exclude_also = ["if TYPE_CHECKING:"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
FAIL_INVALID_TEMPLATE_VARS = true
